requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
djangorestframework-simplejwt==5.3.0
django-celery-results==2.5.1
celery==5.3.4
redis==5.0.1
psycopg2==2.9.7
Pillow==10.0.1
requests==2.31.0
python-decouple==3.8
drf-yasg==1.21.7
whitenoise==6.6.0
gunicorn==21.2.0
python-dotenv==1.0.0